Fix Prepared Statements behaviours after 15252
Patch by Alex Petrov; reviewed by Marcus Eriksson for CASSANDR-17248.

+++ src/java/org/apache/cassandra/gms/Gossiper.java
@@ -1908,7 +1908,7 @@ public class Gossiper implements IFailureDetectionEventListener, GossiperMBean
     }
 
     @Nullable
-    public CassandraVersion getMinVersion(int delay, TimeUnit timeUnit)
+    public CassandraVersion getMinVersion(long delay, TimeUnit timeUnit)
     {
         try
         {
